reflection: move builder and state to parent module
